+++ app/featureset/featureset_internal_test.go
@@ -25,7 +25,6 @@ func TestAllFeatureStatus(t *testing.T) {
 	// Add all features to this test
 	features := []Feature{
 		QBFTConsensus,
-		FrostDKG,
 	}
 
 	for _, feature := range features {
