+++ dkg/dkg.go
@@ -38,14 +38,6 @@ import (
 	"github.com/obolnetwork/charon/tbls/tblsconv"
 )
 
-// Values of following constants should not change as it can break backwards compatibility.
-const (
-	// dutyLock is responsible for lock hash signed partial signatures exchange and aggregation.
-	dutyLock core.DutyType = 101
-	// dutyDepositData is responsible for deposit data signed partial signatures exchange and aggregation.
-	dutyDepositData core.DutyType = 102
-)
-
 type Config struct {
 	DefFile string
 	DataDir string
@@ -54,86 +46,6 @@ type Config struct {
 
 	TestDef     *cluster.Definition
 	TestSigning bool
-}
-
-// sigData includes the fields obtained from sigdb when threshold is reached.
-type sigData struct {
-	duty   core.DutyType
-	pubkey core.PubKey
-	psigs  []core.ParSignedData
-}
-
-// exchanger is responsible for exchanging partial signatures between peers on libp2p.
-type exchanger struct {
-	sigChan chan sigData
-	sigex   *parsigex.ParSigEx
-	sigdb   *parsigdb.MemDB
-	numVals int
-}
-
-func newExchanger(tcpNode host.Host, peerIdx int, peers []peer.ID, vals int) *exchanger {
-	ex := &exchanger{
-		// threshold is len(peers) to wait until we get all the partial sigs from all the peers per DV
-		sigdb:   parsigdb.NewMemDB(len(peers)),
-		sigex:   parsigex.NewParSigEx(tcpNode, peerIdx, peers),
-		sigChan: make(chan sigData),
-		numVals: vals,
-	}
-
-	// Wiring core workflow components
-	ex.sigdb.SubscribeInternal(ex.sigex.Broadcast)
-	ex.sigdb.SubscribeThreshold(ex.pushPsigs)
-	ex.sigex.Subscribe(ex.sigdb.StoreExternal)
-
-	return ex
-}
-
-// exchange exhanges partial signatures of lockhash/deposit-data among dkg participants and returns all the partial
-// signatures of the group according to public key of each DV.
-func (e *exchanger) exchange(ctx context.Context, duty core.DutyType, set core.ParSignedDataSet) (map[core.PubKey][]core.ParSignedData, error) {
-	// Start the process by storing current peer's ParSignedDataSet
-	err := e.sigdb.StoreInternal(ctx, core.Duty{Type: duty}, set)
-	if err != nil {
-		return nil, err
-	}
-
-	sets := make(map[core.PubKey][]core.ParSignedData)
-	for {
-		select {
-		case <-ctx.Done():
-			return nil, ctx.Err()
-		case peerSet := <-e.sigChan:
-			if duty != peerSet.duty {
-				// Do nothing if duty doesn't match
-				continue
-			}
-
-			switch peerSet.duty {
-			case dutyLock, dutyDepositData:
-				sets[peerSet.pubkey] = peerSet.psigs
-			default:
-				return nil, errors.New("invalid data")
-			}
-		}
-
-		// We are done when we have ParSignedData of all the DVs from all each peer
-		if len(sets) == e.numVals {
-			break
-		}
-	}
-
-	return sets, nil
-}
-
-// pushPsigs is responsible for writing partial signature data to sigChan obtained from other peers.
-func (e *exchanger) pushPsigs(_ context.Context, duty core.Duty, pk core.PubKey, psigs []core.ParSignedData) error {
-	e.sigChan <- sigData{
-		duty:   duty.Type,
-		pubkey: pk,
-		psigs:  psigs,
-	}
-
-	return nil
 }
 
 // Run executes a dkg ceremony and writes secret share keystore and cluster lock files as output.
@@ -170,8 +82,6 @@ func Run(ctx context.Context, conf Config) error {
 		return errors.Wrap(err, "hash definition")
 	}
 	clusterID := fmt.Sprintf("%x", defHash[:])
-
-	_ = newExchanger(tcpNode, nodeIdx.PeerIdx, nil, def.NumValidators)
 
 	var shares []share
 	switch def.DKGAlgorithm {
